Digimon Immunity as a Permanent Effect, instead of card effect
